Move common functions to fix import tree
