+++ jobrunner/create_or_update_jobs.py
@@ -8,7 +8,6 @@ and doing the necessary dependency resolution.
 import logging
 from pathlib import Path
 import re
-from urllib.parse import urlparse
 import time
 
 from . import config
@@ -23,9 +22,13 @@ from .database import (
 from .git import (
     read_file_from_repo,
     get_sha_from_remote_ref,
-    commit_reachable_from_ref,
     GitError,
     GitFileNotFoundError,
+)
+from .github_validators import (
+    validate_branch_and_commit,
+    validate_repo_url,
+    GithubValidationError,
 )
 from .project import (
     parse_and_validate_project_file,
@@ -64,7 +67,12 @@ def create_or_update_jobs(job_request):
             log.info(f"Handling new JobRequest:\n{job_request}")
             new_job_count = create_jobs(job_request)
             log.info(f"Created {new_job_count} new jobs")
-        except (GitError, ProjectValidationError, JobRequestError) as e:
+        except (
+            GitError,
+            GithubValidationError,
+            ProjectValidationError,
+            JobRequestError,
+        ) as e:
             log.info(f"JobRequest failed:\n{e}")
             create_failed_job(job_request, e)
         except Exception:
@@ -261,53 +269,6 @@ def validate_job_request(job_request):
         validate_branch_and_commit(
             job_request.repo_url, job_request.commit, job_request.branch
         )
-
-
-def validate_repo_url(repo_url, allowed_gitub_orgs):
-    parsed_url = urlparse(repo_url)
-    if parsed_url.scheme != "https" or parsed_url.netloc != "github.com":
-        raise JobRequestError("Repository URLs must start https://github.com")
-    path = parsed_url.path.strip("/").split("/")
-    if not path or path[0] not in allowed_gitub_orgs:
-        raise JobRequestError(
-            f"Repositories must belong to one of the following Github "
-            f"organisations: {' '.join(allowed_gitub_orgs)}"
-        )
-    expected_url = f"https://github.com/{'/'.join(path[:2])}"
-    if repo_url.rstrip("/") != expected_url or len(path) != 2:
-        raise JobRequestError(
-            "Repository URL was not of the expected format: "
-            "https://github.com/[organisation]/[project-name]"
-        )
-
-
-def validate_branch_and_commit(repo_url, commit, branch):
-    """
-    Due to the way Github works, anyone who can open a pull request against a
-    repository can make a commit appear to be "in" that repository, even if
-    they do not have write access to it.
-
-    For example, someone created this PR against the Linux kernel:
-    https://github.com/torvalds/linux/pull/437
-
-    And even though this will never be merged, it still appears as a commit in
-    that repo:
-    https://github.com/torvalds/linux/commit/2793ae1df012c7c3f13ea5c0f0adb99017999c3b
-
-    If we are enforcing that only code from certain organisations can be run
-    then we need to check that any commits supplied have been made by someone
-    with write access to the repository, which means we need to check they
-    belong to a branch or tag in the repository.
-    """
-    if not branch:
-        raise JobRequestError("A branch name must be supplied")
-    # A further wrinkle is that each PR gets an associated ref within the repo
-    # of the form `pull/PR_NUMBER/head`. So we enforce that the branch name
-    # must be a "plain vanilla" branch name with no slashes.
-    if "/" in branch:
-        raise JobRequestError(f"Branch name must not contain slashes: {branch}")
-    if not commit_reachable_from_ref(repo_url, commit, branch):
-        raise JobRequestError(f"Could not find commit on branch '{branch}': {commit}")
 
 
 def create_failed_job(job_request, exception):
